refactor: 체스 게임 컨트롤러 분리
