+++ src/main/Readme.md
@@ -11,6 +11,7 @@
       - [x] 비밀번호를 올바르게 입력하면 체스방이 삭제된다.
 - [x] 기존 체스방 생성, 게임 수행, 게임 종료의 주소에 고유값을 추가한다.  
 - [ ] 예외가 발생했을 때 사용자가 이해할 수 있는 명시적인 메시지를 응답한다.
+- [ ] 진행 중인 체스방은 삭제할 수 없다.
   
 
 ### 추가로 생각해볼 부분
